% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotly_contributions.R
\name{ClusterComposition}
\alias{ClusterComposition}
\title{ClusterComposition}
\usage{
ClusterComposition(data, clustering = "edited_res.1.5", condition = "stage")
}
\arguments{
\item{data}{metadata to be used}

\item{clustering}{clustering column used for seurat clusters}

\item{condition}{which data should be used for plotting}
}
\value{
plotly object
}
\description{
ClusterComposition
}
\examples{
ClusterComposition(data = metadata_all, clustering = "edited_res.1.5, condition = "stage)
}
