% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{compare_expression}
\alias{compare_expression}
\title{Compare Expression Between Two Cell Types}
\usage{
compare_expression(data, gene, type1, type2, method = "t.test")
}
\arguments{
\item{data}{An object of class `scData`.}

\item{gene}{A character string for the gene to test.}

\item{type1}{The first cell type group.}

\item{type2}{The second cell type group.}

\item{method}{The statistical test ("t.test" or "wilcox"). Default is "t.test".}
}
\value{
An object of class `htest` containing the formal test results and p-value.
}
\description{
Performs a two-sample t-test or Wilcoxon test to mathematically determine 
  if a gene is "differentially expressed" (significantly altered) between two cell types.
}
\examples{
data <- load_scrna()
compare_expression(data, "CD79A", "B cell", "CD4+ T cell")
}
