% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{custom_DimPlot}
\alias{custom_DimPlot}
\title{Title}
\usage{
custom_DimPlot(
  seurat,
  metadata_clusters = "custom_int_res2",
  selected_label = c("label", "new_label"),
  wrap_length = 10
)
}
\arguments{
\item{seurat}{seurat object}

\item{metadata_clusters}{metadata to choose for clustering}
}
\value{
ggplot object
}
\description{
Title
}
