% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotly_plots.R
\name{empty_plot}
\alias{empty_plot}
\title{Make empty ggplot2 or plotly}
\usage{
empty_plot(label = " ", return_plotly = FALSE)
}
\arguments{
\item{label}{text to be displayed}

\item{return_plotly}{if return plotly or ggplot2 object}
}
\value{
ggplot2 or plotly object
}
\description{
Make empty ggplot2 or plotly
}
