% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment_analysis.R
\name{enrichment_analysis_UI}
\alias{enrichment_analysis_UI}
\title{enrichment_analysis_UI}
\usage{
enrichment_analysis_UI(
  id,
  label = "Paste input genes here: ",
  info = "Upload your genes of interest for gene set enrichment analysis",
  placeholder = "myh6\\ncmlc2\\nisl1\\ntbx2a\\n..."
)
}
\description{
enrichment_analysis_UI
}
