% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expression_level.R
\name{ExpressionLevelShiny}
\alias{ExpressionLevelShiny}
\title{ExpressionLevelShiny}
\usage{
ExpressionLevelShiny(
  id,
  clustering,
  selected = "myh6",
  metadata = NULL,
  slot_data = NULL
)
}
\description{
ExpressionLevelShiny
}
