% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expression_level.R
\name{ExpressionLevelUI}
\alias{ExpressionLevelUI}
\title{FeaturePlotShiny UI}
\usage{
ExpressionLevelUI(
  id,
  gene_label = "Select gene",
  cluster_label = "Cluster:",
  selected = "myh6",
  choices = levels(metadata_all$edited_res.1.5),
  width = "100px"
)
}
\description{
FeaturePlotShiny UI
}
