% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_plot.R
\name{FeaturePlotShiny}
\alias{FeaturePlotShiny}
\title{FeaturePlotShiny}
\usage{
FeaturePlotShiny(
  id,
  metadata = NULL,
  data_slot = NULL,
  identity,
  order = FALSE,
  stage,
  selected = "myh6",
  pt.size = 5e-04,
  pt.size_pos = 0.5
)
}
\description{
FeaturePlotShiny
}
