% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_plot.R
\name{FeaturePlotShinyUI}
\alias{FeaturePlotShinyUI}
\title{FeaturePlotShiny UI}
\usage{
FeaturePlotShinyUI(
  id,
  label = "Choose gene",
  value = "",
  placeholder = "",
  width = "100px",
  spinnertype = 5
)
}
\description{
FeaturePlotShiny UI
}
