% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-features_in_assay.R
\name{features_in_assay}
\alias{features_in_assay}
\title{Get names of all features in an assay/experiment}
\usage{
features_in_assay(object, assay, ...)
}
\arguments{
\item{object}{a single-cell object.}

\item{assay}{the name of an assay/modality for which to view features}

\item{...}{Currently unused.}
}
\description{
Returns the names of all features in an assay/modality
}
\keyword{internal}
