% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Contributions.R
\name{get_data_metrics}
\alias{get_data_metrics}
\title{Get metrics for feature}
\usage{
get_data_metrics(
  metadata = metadata_all,
  main_group = "edited_res.1.5",
  feature,
  split_by = NULL
)
}
\arguments{
\item{metadata}{data to use}

\item{main_group}{cluster order to use}

\item{feature}{metric to use}

\item{split_by}{option if want to split by that feature}
}
\value{
summarized grouped data
}
\description{
Get metrics for feature
}
