% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scData.R
\name{load_scrna}
\alias{load_scrna}
\title{Load scRNA Dataset}
\usage{
load_scrna()
}
\value{
An object of class `scData`, packing the sparse expression, PCA, and cell labels.
}
\description{
Loads the actual 10X Genomics PBMC dataset and assigns it an S3 class 
  called `scData` so we can easily use Object-Oriented methods on it.
  Single-cell RNA sequencing (scRNA-seq) lets us observe gene expression in individual cells.
}
\examples{
data <- load_scrna()
print(data)
}
