% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_subset.R
\name{make_subset}
\alias{make_subset}
\alias{make_subset.default}
\alias{make_subset.AnnDataR6}
\title{Subset Creation}
\usage{
make_subset(object, filter_list)

\method{make_subset}{default}(object, filter_list)

\method{make_subset}{AnnDataR6}(object, filter_list)
}
\arguments{
\item{object}{a single cell object}

\item{filter_list}{a list of filters constructed by the subset_selections module.}
}
\value{
A Seurat object subsetted for the criteria entered.
}
\description{
Subset Creation

Method ran for all classes besides anndata
}
\section{Methods (by class)}{
\itemize{
\item \code{make_subset(default)}: Seurat, SingleCellExperiment objects

\item \code{make_subset(AnnDataR6)}: Anndata objects

}}
\keyword{internal}
