% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{markers2Csv}
\alias{markers2Csv}
\title{Function save markers to excel file, each cluster on separate sheet}
\usage{
markers2Csv(markers, file, outdir)
}
\arguments{
\item{markers}{seuratFindMarkers output}

\item{file}{A character string naming an xlsx file}

\item{outdir}{A character path to outdir}
}
\value{
Return a xlsx file
}
\description{
Function save markers to excel file, each cluster on separate sheet
}
\examples{
 
markers2Csv(markers = markers, file = "markers.xlsx", outdir = "/path/to/directory/")
}
