% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Contributions.R
\name{metadata_prop}
\alias{metadata_prop}
\title{metadata_prop}
\usage{
metadata_prop(metadata, main_group, feature, stage = NULL)
}
\arguments{
\item{metadata}{metadata}

\item{main_group}{clustering based on which metadata}

\item{feature}{feature to use}

\item{stage}{needed for cell cycle plot across stage}
}
\value{
ggplot barplot
}
\description{
metadata_prop
}
