% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Contributions.R
\name{MultiPlot_UI}
\alias{MultiPlot_UI}
\title{MultiPlot_UI}
\usage{
MultiPlot_UI(
  id,
  label = "Metadata:",
  selected = "Overview",
  choices = c("Overview", Dataset = "DataSet", Line = "line", Stage = "stage",
    "Percent.mt", "Percent.hemoglobin", `Cell cycle` = "Phase", `Cell cycle-timepoint` =
    "Phase_timepoint", `Number of genes` = "nFeature_RNA", `Number of UMI` = "log10_UMI",
    Iris = "iris")
)
}
\arguments{
\item{id}{}

\item{label}{}

\item{selected}{}

\item{choices}{}
}
\description{
MultiPlot_UI
}
