% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_FeaturePlot.R
\name{my_FeaturePlot}
\alias{my_FeaturePlot}
\title{Custom FeaturePlot}
\usage{
my_FeaturePlot(
  metadata,
  data_slot,
  gene,
  identity,
  pt.size = 5e-04,
  pt.size_pos = 0.005,
  label = FALSE,
  order = FALSE,
  exp_lvl = 0
)
}
\arguments{
\item{metadata}{metadata exported from seurat}

\item{data_slot}{slot "data" from seurat}

\item{gene}{gene to be plotted}

\item{identity}{}

\item{pt.size}{dot size}

\item{pt.size_pos}{dot size of cells to be colored}

\item{label}{to plot labels or not}

\item{order}{whether sort plotting order or not}

\item{exp_lvl}{gene expression level to use as a threshold}
}
\description{
Custom FeaturePlot
}
