% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncbi2gene.R
\name{ncbi2gene}
\alias{ncbi2gene}
\title{ncbi2gene}
\usage{
ncbi2gene(enrichResult, gene_id_type = NULL)
}
\arguments{
\item{enrichResult}{}

\item{gene_name_type}{ensembl or extrenal gene name}
}
\value{

}
\description{
ncbi2gene
}
