% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scData.R
\name{plot.scData}
\alias{plot.scData}
\title{Plot method for scData}
\usage{
\method{plot}{scData}(x, ...)
}
\arguments{
\item{x}{An object of class `scData`}

\item{...}{Additional arguments passed to `plot_pca`}
}
\description{
A generic plot wrapper that defaults to PCA visualization to show biological clustering.
}
