% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotly_plots.R
\name{plotly_cc_stage}
\alias{plotly_cc_stage}
\title{Cell-cycle bar chart in plotly}
\usage{
plotly_cc_stage(
  data = metadata_all,
  identity = "edited_res.1.5",
  stage = "stage",
  feature = "Phase",
  my_colors = rev(c("#0B775E", "#DC863B"))
)
}
\arguments{
\item{data}{metadata}

\item{identity}{main clustering group}

\item{stage}{condition for grouping}

\item{feature}{condition to plot}

\item{my_colors}{colors}
}
\value{

}
\description{
Cell-cycle bar chart in plotly
}
