% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw_ngene_mt.R
\name{raw_ngene_mt}
\alias{raw_ngene_mt}
\title{Reading raw metadata for all cells}
\usage{
raw_ngene_mt(input_metadata = overview_metadata)
}
\arguments{
\item{input_metadata}{overview_metadata loaded from all data}
}
\value{
... ggplot object
}
\description{
Reading raw metadata for all cells
}
