% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_multiple_sheets.R, R/utils.R
\name{read_multiple_sheets}
\alias{read_multiple_sheets}
\title{Read multiple excel sheets}
\usage{
read_multiple_sheets(file)

read_multiple_sheets(file)
}
\arguments{
\item{file}{path to the file}
}
\value{
a data.frame
}
\description{
Read multiple excel sheets

Read multiple excel sheets
}
