% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{run_anova}
\alias{run_anova}
\title{Run One-Way ANOVA on Marker Genes}
\usage{
run_anova(data, gene)
}
\arguments{
\item{data}{An object of class `scData`.}

\item{gene}{A character string specifying the gene to test.}
}
\value{
An object of class `aov`.
}
\description{
Implements a One-Way ANOVA to test if the mean expression of a gene is 
  statistically identical across all cell types simultaneously, or if at least one cluster differs.
}
\examples{
data <- load_scrna()
run_anova(data, "CD8A")
}
