% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_scExploreR.R
\name{run_scExploreR}
\alias{run_scExploreR}
\title{scExploreR app}
\usage{
run_scExploreR(port = NULL)
}
\arguments{
\item{port}{specify a port for launching the browser. This is required to
run several instances of the browser at the same IP address. See "how to run scExploreR" for more information.}
}
\description{
Initializes the main scExploreR app.
}
\examples{
run_scExploreR(port = 5320)
}
