% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotly_plots.R
\name{stacked_bar_plotly}
\alias{stacked_bar_plotly}
\title{Stacked plotly bar plot}
\usage{
stacked_bar_plotly(
  metadata = metadata_all,
  feature,
  main_group = "edited_res.1.5"
)
}
\arguments{
\item{metadata}{metadata}

\item{feature}{feature}

\item{main_group}{metric}
}
\value{
plotly object
}
\description{
Stacked plotly bar plot
}
