% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotly_contributions.R
\name{StackedMetrics}
\alias{StackedMetrics}
\title{Title}
\usage{
StackedMetrics(
  data,
  clustering = "edited_res.1.5",
  condition = c("nCount_RNA", "nFeature_RNA", "percent.mt")
)
}
\arguments{
\item{data}{metadata to be used}

\item{clustering}{clustering column used for seurat clusters}

\item{condition}{which data should be used for plotting: nCount_RNA, n"Feature_RNA, percent.mt}
}
\value{
plotly plot
}
\description{
Title
}
