% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotly_plots.R
\name{violin_metadata_stage}
\alias{violin_metadata_stage}
\title{Violin plot between one metadata split by stage}
\usage{
violin_metadata_stage(
  metadata = metadata_all,
  feature,
  main_group = "edited_res.1.5",
  split_by = "stage"
)
}
\arguments{
\item{metadata}{metadata}

\item{feature}{feature}

\item{main_group}{metric}

\item{split_by}{split plot by}
}
\value{
plotly object
}
\description{
Violin plot between one metadata split by stage
}
