% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotly_plots.R
\name{violin_plotly}
\alias{violin_plotly}
\title{Violin_plotly}
\usage{
violin_plotly(metadata = metadata_all, CLUSTERS = NULL)
}
\arguments{
\item{metadata}{}

\item{cluster}{}
}
\description{
Violin_plotly
}
