% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volcano_in_cluster.R
\name{volcano_plot}
\alias{volcano_plot}
\title{Volcano plot}
\usage{
volcano_plot(
  seurat_object,
  markers = NULL,
  ident.1 = NULL,
  ident.2 = NULL,
  avg_log2FC.1 = NULL,
  avg_log2FC.2 = NULL,
  plot_pval = 1e-296,
  height = 15,
  label.ident.1 = ident.1,
  label.ident.2 = ident.2,
  pos.label.1 = pos_label_1,
  pos.label.2 = pos_label_2,
  label.title.size = 5.5,
  ann_text_size = 4.5,
  plot_top = FALSE,
  n_genes = 5,
  point.size = 3,
  alpha = 0.6
)
}
\arguments{
\item{seurat_object}{seurat object}

\item{markers}{output from FindMarkers/FindAllMarkers from Seurat

importFrom ggplot2 aes}

\item{ident.1}{}

\item{ident.2}{}

\item{avg_log2FC.1}{}

\item{avg_log2FC.2}{}

\item{plot_pval}{}

\item{height}{}

\item{label.ident.1}{}

\item{label.ident.2}{}

\item{pos.label.1}{}

\item{pos.label.2}{}

\item{label.title.size}{}

\item{ann_text_size}{}

\item{plot_top}{}

\item{n_genes}{}

\item{point.size}{}

\item{alpha}{}
}
\description{
Volcano plot
}
