% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volcano_in_cluster.R
\name{volcano_plot_in_cluster}
\alias{volcano_plot_in_cluster}
\title{Volcano plot side by side in particular clusters}
\usage{
volcano_plot_in_cluster(
  markers = NULL,
  seurat,
  clusters = NULL,
  n_pos = 8,
  n_neg = -8,
  dot_size = 3,
  order.by = c("avg_log2FC", "p_val_adj"),
  label.size = 4.5
)
}
\arguments{
\item{markers}{#positive and negative markers}

\item{seurat}{#seurat object}

\item{clusters}{#clusters to be plotted}

\item{n_pos}{#how many positive genes to annotate}

\item{n_neg}{#how many negative genes to annotate}

\item{dot_size}{#size of the dot}

\item{order.by}{}

\item{label.size}{}
}
\description{
Volcano plot side by side in particular clusters
}
