% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_plot.R
\name{FeaturePlotShinyUI}
\alias{FeaturePlotShinyUI}
\title{FeaturePlotShiny UI}
\usage{
FeaturePlotShinyUI(
  id,
  label = "Select gene:",
  value = "",
  placeholder = "myh6",
  width = "100px",
  label_radio = "Stage:"
)
}
\description{
FeaturePlotShiny UI
}
