% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_config.R, R/generate_config_yaml.R
\name{generate_config_yaml}
\alias{generate_config_yaml}
\title{Auto-Generate Config File}
\usage{
generate_config_yaml(
 # Replace with path to your object
 object = path_to_your_object,
 file = output_config_path,
 object_name = "Test Single-Cell Object",
 object_description = "This is a test object",
 genes_assay = "RNA",
 include_numeric_metadata = TRUE
 # Set the
 is_HDF5SummarizedExperiment = TRUE
 )

auto_config(
 # Replace with path to your object
 object = path_to_your_object,
 file = output_config_path,
 object_name = "Test Single-Cell Object",
 object_description = "This is a test object",
 genes_assay = "RNA",
 include_numeric_metadata = TRUE
 # Set the
 is_HDF5SummarizedExperiment = TRUE
 )
}
\arguments{
\item{object}{path to a single-cell object to be configured. Currently,
Seurat, SingleCellExperiment, and anndata objects are supported. For
SingleCellExperiment objects with using HDF5 disk-backed storage via
\link{HDF5Array}, \code{object_path} should be a path to the directory containing
the se.rds and assays.h5 files for the object.}

\item{file}{path specifying where the output config file should be saved.
This should end in ".yaml".}

\item{object_name}{a single-length character vector with the display name
of the object. This will be displayed to end users in the app as entered.}

\item{object_description}{a single-length character vector with a short
description of the object. This will be displayed to end users in the
dataset selection menu. This should be a few sentences long, and include
object details relevant to end-users (i.e. brief descripton of methods,
number of cells/samples in object, etc.).}

\item{is_HDF5SummarizedExperiment}{Set this to TRUE to load an HDF5-enabled
SingleCellExperiment object saved via saveHDF5SummarizedExperiment. When
loading an HDF5-enabled object, set the object_path to the directory of the
HDF5-enabled object, created when saving the object via
HDF5Array:saveHDF5SummarizedExperiment.}

\item{HDF5_prefix}{When loading an HDF5-backed SingleCellExperiment HDF5_prefix
is passed to the prefix parameter of HDF5Array::loadHDF5SummarizedExperiment
to specify the prefixes for the se.rds and assays.h5 files.}

\item{include_numeric_metadata}{If \code{TRUE}, numeric metadata in the object
will searchable in scExploreR for plotting and subsetting. The default
setting is \code{TRUE}.}

\item{genes_assay}{single-length character vector giving the name of the
genes assay/experiment/modality in the object. If undefined, the first assay
in the object will be used. The name should be entered using the following
convention:
\itemize{
\item Seurat objects: use the name of the assay as it appears in \code{names(object@assays)}.
\item SingleCellExperiment objects: use the name of the experiment as it appears in \code{\link[SingleCellExperiment:altExps]{SingleCellExperiment::altExpNames()}} or \code{\link[SingleCellExperiment:altExps]{SingleCellExperiment::mainExpName()}}
\item anndata objects: use either "X" to specify the modality measured in X, or enter the name of a matrix stored in the \code{obsm} slot.
}}

\item{sample_level_var}{a sample-level metadata variable used to construct
sample-level pie charts in the plots tab. This should be a sample or patient
ID, for example. If this is not provided (the default), pie charts will not
appear in scExploreR.}
}
\value{
A config file is generated at the path provided to \code{file}. Nothing is returned from the function.
}
\description{
For objects with many metadata variables, setting up a config file in the
config app GUI may be time-consuming. To facilitate deployment of such
objects, this function automatically generates a config file with all assays,
metadata, and reductions enabled, using the names for each as they appear
in the object. To customize settings for the object, the YAML file produced
by this function can be edited by hand, or loaded into the config app via \code{run_config_app()}.
}
\details{
The structure of the config file generated is as follows:
\itemize{
\item All assays/experiments/modalities in the object are included, and the
display names for each are set to the names of each as they appear in the
object. The "suffixes" with the display name of the assay that appear on
plots to indicate the assay of origin for features will not display. This
must be set using the config app, or manually via the \code{suffix_human} key.
\item All categorical metadata variables are included. Numeric metadata variables
can't be customized in config files in the current version of scExploreR.
Numeric metadata may be exposed in the app for plotting and subsetting by
setting the \code{include_numeric_metadata} parameter to \code{TRUE}. Metadata variables
are displayed in the order they were entered in the object metadata table.
\item All reductions are included and are named as they appear in the object.
}
}
