% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Contributions.R
\name{MultiPlot_UI}
\alias{MultiPlot_UI}
\title{MultiPlot_UI}
\usage{
MultiPlot_UI(
  id,
  label = "Type of data",
  selected = "Overview",
  choices = c("Overview", "Dataset", "Line", "Stage", "Percent.mt",
    "Percent.hemoglobin", "Cell-Cycle-phase", "nCount_RNA", "nFeature_RNA", "iris")
)
}
\arguments{
\item{id}{}

\item{label}{}

\item{selected}{}

\item{choices}{}
}
\description{
MultiPlot_UI
}
