% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Contributions.R
\name{MultiPlot_UI}
\alias{MultiPlot_UI}
\title{MultiPlot_UI}
\usage{
MultiPlot_UI(
  id,
  label = "Select category:",
  selected = "selected_cells",
  choices = c(`Selected cells` = "selected_cells", Overview = "Overview", Contribution =
    "contribution", `Cluster metrics` = "qc_cluster", `Global metrics` = "qc"),
  width = "100\%"
)
}
\arguments{
\item{id}{}

\item{label}{}

\item{selected}{}

\item{choices}{}
}
\description{
MultiPlot_UI
}
