% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Contributions.R
\name{MultiPlot_UI}
\alias{MultiPlot_UI}
\title{MultiPlot_UI}
\usage{
MultiPlot_UI(
  id,
  label = "Metadata:",
  selected = "Overview",
  choices = c(Overview = "Overview", Contribution = "contribution", `General QC` =
    "qc", `QC in cluster` = "qc_cluster")
)
}
\arguments{
\item{id}{}

\item{label}{}

\item{selected}{}

\item{choices}{}
}
\description{
MultiPlot_UI
}
