% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_FeaturePlot.R, R/notatnik.R
\name{my_FeaturePlot}
\alias{my_FeaturePlot}
\title{Custom FeaturePlot}
\usage{
my_FeaturePlot(
  metadata,
  data_slot,
  gene,
  identity,
  label = FALSE,
  order = FALSE
)

my_FeaturePlot(
  metadata,
  data_slot,
  gene,
  identity,
  label = FALSE,
  order = FALSE
)
}
\arguments{
\item{metadata}{metadata exported from seurat}

\item{data_slot}{slot "data" from seurat}

\item{gene}{gene to be plotted}

\item{label}{to plot labels or not}

\item{order}{whether sort plotting order or not}

\item{identitty}{identity of interest (Idents(seurat))}
}
\description{
Custom FeaturePlot

Custom FeaturePlot
}
