% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Contributions.R
\name{plot_contribution}
\alias{plot_contribution}
\title{Plot contributions}
\usage{
plot_contribution(
  metadata,
  main_group = "edited_res.1.5",
  feature = c("line", "stage", "DataSet", "iris", "Phase", "Phase_timepoint",
    "nFeature_RNA", "percent.mt"),
  choice
)
}
\arguments{
\item{metadata}{}

\item{main_group}{}

\item{feature}{}

\item{choice}{}
}
\description{
Plot contributions
}
