% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw_ngene_mt.R
\name{raw_ngene_mt}
\alias{raw_ngene_mt}
\title{Function description... TODO}
\usage{
raw_ngene_mt(input_metadata)
}
\arguments{
\item{input_metadata}{... TODO}
}
\value{
... TODO ggplot object?
}
\description{
Function description... TODO
}
