% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_config_app.R
\name{run_config}
\alias{run_config}
\title{scExploreR config app}
\usage{
run_config(./path_to_object.rds, ./path_to_config_file.yaml)
}
\arguments{
\item{object_path}{relative or absolute path to the Seurat object to be
configured. The file provided must be a .rds file created from a single
Seurat object.}

\item{config_path}{optional: if provided, the data from this file will be 
loaded when the user selects "load config file" in the config app. This 
should be a YAML file, though .rds files from versions 0.4.0 and earlier will
also be accepted.}

\item{is_HDF5SummarizedExperiment}{Set this to TRUE to load an HDF5-enabled 
SingleCellExperiment object saved via saveHDF5SummarizedExperiment.}

\item{dev_mode}{Used only for development. If TRUE, the server values for each option chosen by the user will be printed at the bottom of the "general" tab.}
}
\description{
run_config() will launch a Shiny app used to configure datasets for use in 
the main browser.
}
