% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_scExploreR.R
\name{run_scExploreR}
\alias{run_scExploreR}
\title{scExploreR app}
\usage{
run_scExploreR(browser_config, port = NULL)
}
\arguments{
\item{browser_config}{path to a YAML config file giving browser specific settings. See LINK for more info on creating this file.}

\item{port}{specify a port for launching the browser. This is required to run several instances of the browser at the same IP address. See "how to run scExploreR" for more information. The port can be any number between 3000:8000, except for ports blocked by Google Chrome (for more information on this, see \link[shiny]{runApp}).}
}
\description{
Initializes the main scExploreR app.
}
